fix: set correct seed peer features
